api/requirements.txt
Flask==3.1.0
numpy==2.2.1
Pillow==11.1.0
torch==2.5.1
torchvision==0.20.1
